+++ test/golden_test.dart
@@ -23,4 +23,17 @@ void main() {
     expect(script, contains('complete -c my_app -f -a'));
     expect(script, contains('###-end-my_app-completion-###'));
   });
+
+  test('generate script for nushell', () {
+    final script = generateCompletionScript(['my_app'], shell: Shell.nushell);
+
+    expect(script, contains('###-begin-my_app-completion-###'));
+    expect(script, isNot(contains('###-for-bash-###')));
+    expect(script, isNot(contains('###-for-zsh-###')));
+    expect(script, isNot(contains('###-for-fish-###')));
+    expect(script, contains('###-for-nushell-###'));
+    expect(script, contains('def "__my_app_completion"'));
+    expect(script, contains('export extern "my_app"'));
+    expect(script, contains('###-end-my_app-completion-###'));
+  });
 }